pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
authors = ["Artem <serov4rtem@mail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
bs4 = "^0.0.2"
validators = "^0.33.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
